set BackgroundInfo as generic data attribute in kaleido parser
